+++ Gateway/Response/ResponseHandler.php
@@ -30,23 +30,9 @@ use Alma\MonthlyPayments\Gateway\Config\Config;
 use Magento\Payment\Gateway\Helper\SubjectReader;
 use Magento\Payment\Gateway\Response\HandlerInterface;
 use Magento\Sales\Model\Order;
-use Alma\MonthlyPayments\Helpers\Logger;
 
 class ResponseHandler implements HandlerInterface
 {
-    /**
-     * @var Logger
-     */
-    private $logger;
-
-    /**
-     * @param Logger $logger
-     */
-    public function __construct(Logger $logger)
-    {
-        $this->logger = $logger;
-    }
-
     /**
      * Handles transaction id
      *