+++ Model/Adminhtml/Config/PaymentPlans.php
@@ -25,9 +25,7 @@
 
 namespace Alma\MonthlyPayments\Model\Adminhtml\Config;
 
-use Alma\MonthlyPayments\Gateway\Config\Config;
 use Alma\MonthlyPayments\Gateway\Config\PaymentPlans\PaymentPlanConfig;
-use Alma\MonthlyPayments\Gateway\Config\PaymentPlans\PaymentPlansConfig;
 use Alma\MonthlyPayments\Gateway\Config\PaymentPlans\PaymentPlansConfigInterfaceFactory;
 use Magento\Config\Model\Config\Backend\Serialized;
 use Magento\Framework\App\Cache\TypeListInterface;
@@ -53,10 +51,6 @@ class PaymentPlans extends Serialized
      */
     protected $hasError;
     /**
-     * @var Config
-     */
-    private $almaConfig;
-    /**
      * @var Json
      */
     protected $serializer;
@@ -64,6 +58,10 @@ class PaymentPlans extends Serialized
      * @var PaymentPlansConfigInterfaceFactory
      */
     private $plansConfigFactory;
+    /**
+     * @var Logger
+     */
+    private $logger;
 
     /**
      * @param Context $context
@@ -72,7 +70,6 @@ class PaymentPlans extends Serialized
      * @param TypeListInterface $cacheTypeList
      * @param MessageManager $messageManager
      * @param PaymentPlansConfigInterfaceFactory $plansConfigFactory
-     * @param Config $almaConfig
      * @param AbstractResource|null $resource
      * @param AbstractDb|null $resourceCollection
      * @param array $data
@@ -86,7 +83,6 @@ class PaymentPlans extends Serialized
         TypeListInterface $cacheTypeList,
         MessageManager $messageManager,
         PaymentPlansConfigInterfaceFactory $plansConfigFactory,
-        Config $almaConfig,
         Logger $logger,
         AbstractResource $resource = null,
         AbstractDb $resourceCollection = null,
@@ -105,7 +101,6 @@ class PaymentPlans extends Serialized
             $serializer
         );
 
-        $this->almaConfig = $almaConfig;
         $this->messageManager = $messageManager;
         $this->serializer = $serializer ?: new Json();
         $this->plansConfigFactory = $plansConfigFactory;
@@ -143,7 +138,7 @@ class PaymentPlans extends Serialized
         if (!is_array($value)) {
             $value = $this->serializer->unserialize($value);
         }
-        $value = $this->correctMinMaxLimit($value);
+        $value = $this->forceMinMaxLimit($value);
         // Remove transient values from the serialized data: it should always come fresh from the API
         foreach (PaymentPlanConfig::transientKeys() as $key) {
             foreach ($value as $planKey => &$planConfig) {
@@ -158,16 +153,26 @@ class PaymentPlans extends Serialized
     }
 
     /**
-     * @param $plans
+     * Check if min and max value are in Alma limit for each plan configuration before serialize it in core_config database
+     *
+     * @param array $plans
      * @return array
      */
-    private function correctMinMaxLimit($plans):array
+    private function forceMinMaxLimit(array $plans):array
     {
         foreach ($plans as &$plan) {
-            if($plan[PaymentPlanConfig::KEY_MIN_AMOUNT]<$plan[PaymentPlanConfig::TRANSIENT_KEY_MIN_ALLOWED_AMOUNT]){
+            if($plan[PaymentPlanConfig::KEY_MIN_AMOUNT]<$plan[PaymentPlanConfig::TRANSIENT_KEY_MIN_ALLOWED_AMOUNT] || $plan[PaymentPlanConfig::KEY_MIN_AMOUNT]>$plan[PaymentPlanConfig::KEY_MAX_AMOUNT]){
+                $this->logger->info('Change min Amount: '.$plan[PaymentPlanConfig::KEY_MAX_AMOUNT].' - Min allowed Amount: '.$plan[PaymentPlanConfig::TRANSIENT_KEY_MIN_ALLOWED_AMOUNT]. '- Max Amount: '.$plan[PaymentPlanConfig::KEY_MAX_AMOUNT],[]);
+                $this->messageManager->addErrorMessage(
+                    sprintf(__("Minimum amount is %s€ for plan %s"),($plan[PaymentPlanConfig::TRANSIENT_KEY_MIN_ALLOWED_AMOUNT]/100),$plan["key"])
+                );
                 $plan[PaymentPlanConfig::KEY_MIN_AMOUNT] = $plan[PaymentPlanConfig::TRANSIENT_KEY_MIN_ALLOWED_AMOUNT];
             }
-            if($plan[PaymentPlanConfig::KEY_MAX_AMOUNT]>$plan[PaymentPlanConfig::TRANSIENT_KEY_MAX_ALLOWED_AMOUNT]){
+            if($plan[PaymentPlanConfig::KEY_MAX_AMOUNT]>$plan[PaymentPlanConfig::TRANSIENT_KEY_MAX_ALLOWED_AMOUNT] || $plan[PaymentPlanConfig::KEY_MAX_AMOUNT]<$plan[PaymentPlanConfig::KEY_MIN_AMOUNT]){
+                $this->logger->info('Change max Amount: '.$plan[PaymentPlanConfig::KEY_MAX_AMOUNT].' - Max allowed Amount: '.$plan[PaymentPlanConfig::TRANSIENT_KEY_MAX_ALLOWED_AMOUNT]. '- Min Amount: '.$plan[PaymentPlanConfig::KEY_MIN_AMOUNT],[]);
+                $this->messageManager->addErrorMessage(
+                    sprintf(__("Maximum amount is %s€ for plan %s"),($plan[PaymentPlanConfig::TRANSIENT_KEY_MAX_ALLOWED_AMOUNT]/100),$plan["key"])
+                );
                 $plan[PaymentPlanConfig::KEY_MAX_AMOUNT] = $plan[PaymentPlanConfig::TRANSIENT_KEY_MAX_ALLOWED_AMOUNT];
             }
         }